actions/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML == 6.0

nltk == 3.7

numpy == 1.19.5

openai == 0.20.0

pandas == 1.4.3

python_dotenv == 0.20.0

rapidfuzz == 2.1.3

spacy == 3.3.0

textstat == 0.7.2

wordfreq == 3.0.1
